Extract the runloop tests into a separate mod
